% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{append_geocode}
\alias{append_geocode}
\title{Append geocodes to a data frame that includes addresses}
\usage{
append_geocode(df, cols, ...)
}
\arguments{
\item{df}{A \code{data.frame}}

\item{cols}{A vector of names or index numbers of the columns that identify 
the address-related columns in the data. Ideally, these should be in a 
natural order (street, city, state, zip, country).}

\item{...}{Other arguments passed to \code{\link{geocode}}}
}
\value{
The original \code{data.frame} with the added columns \code{address},
\code{lat}, and \code{lon}
}
\description{
Append geocodes to a data frame that includes addresses
}

