% Generated by roxygen2 (4.0.1): do not edit by hand
\name{geocode}
\alias{geocode}
\title{Get geo-codes from addresses}
\usage{
geocode(addresses, src = "google")
}
\arguments{
\item{addresses}{A character vector of addresses}

\item{src}{Which geo-coding API to use? Currently "google" is the only
available option}
}
\value{
A \code{data.frame} with three columns: \code{address},
\code{lat}, and \code{lon}
}
\description{
Get geo-codes from addresses
}

