% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode_google}
\alias{geocode_google}
\title{Geocode addresses using the Google Maps API}
\usage{
geocode_google(addresses)
}
\arguments{
\item{addresses}{A character vector of addresses}
}
\value{
A data frame with three columns: \code{address}, \code{latitude},
and \code{longitude}.
}
\description{
This is \code{geocode} with \code{provider = "google"}. Note that by using this 
function, you're agreeing to the Google Maps API terms of service. See 
\url{https://developers.google.com/maps/terms#section_10_12}.
}
\note{
You will be limited to 2,500 addresses per 24-hour period.
}

