% Generated by roxygen2 (4.0.1): do not edit by hand
\name{geocode_mapquest}
\alias{geocode_mapquest}
\title{Geocode addresses using the Mapquest Open Geocoding API}
\usage{
geocode_mapquest(addresses, key = api_key("mapquest"), batch = TRUE,
  batch_size = 100L)
}
\arguments{
\item{addresses}{A character vector of addresses}

\item{key}{Your mapquest API key}

\item{batch}{Reserved for future use}

\item{batch_size}{Addresses will be geocoded using the batch geocoding
service. The \code{batch_size} controls how large each batch should be.}
}
\description{
You will need an API key to use this function. You can request one from
\url{http://developer.mapquest.com/web/products/open/geocoding-service}.
}

