% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_leaflet}
\alias{make_leaflet}
\title{Make a leaflet map}
\usage{
make_leaflet(map_data, width = 900, height = 600,
  map_center = as.numeric(map_data[1, c("latitude", "longitude")]),
  default_zoom = 11)
}
\arguments{
\item{map_data}{Data frame to be mapped -- at the very least, this should
include the columns \code{latitude} and \code{longitude}. Other data can be
passed to the text popups}

\item{width}{Width in pixels for the map}

\item{height}{height in pixels for the map}

\item{map_center}{A vector witht the latitude and longitude where the map
should be centered when it is first opened}

\item{default_zoom}{The zoom level that the map will be at when first opened}
}
\description{
Make a leaflet map
}

