% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prep_leaf_data}
\alias{prep_leaf_data}
\title{Rearrange geocoded data to a leaflet map-ready format}
\usage{
prep_leaf_data(geo_data, popup_columns = NULL)
}
\arguments{
\item{geo_data}{A data frame that has columns \code{latitude} and
\code{longitude}}

\item{popup_columns}{The names of the columns in the data frame that should
be used to populate the pop-ups on the map. By default, this will include all
columns except for the latitude and longitude columns.}
}
\description{
Rearrange geocoded data to a leaflet map-ready format
}

