% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_leaflet.R
\name{save_leaflet}
\alias{save_leaflet}
\title{Save a map you've already made to an HTML file}
\usage{
save_leaflet(leafmap, location = NULL)
}
\arguments{
\item{leafmap}{The leaflet map you'd like to save}

\item{location}{If you know where you'd like to save the file, enter the 
full path (including the name of the file). If you don't enter a location, 
you'll get a graphical prompt to click your way to a save location.}
}
\description{
Save a leaflet map to a standalone HTML file. This will output a file that 
you can send to clients or upload. By default, you'll be prompted with a 
file window asking where you want to save the file.
}
\examples{
map_data <- append_geocode(urel)
leaf <- make_leaflet(map_data)
save_leaflet(leaf)  ## will bring up a save file dialog window
## alternatively:
save_leaflet(leaf, "my map.html")
}

