% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{urel}
\alias{urel}
\title{UREL places of interest}
\format{A data frame with 3 rows and 6 variables:
\describe{
  \item{names}{person name}
  \item{giving}{lifetime giving}
  \item{street}{street address}
  \item{city}{city name (Berkeley)}
  \item{state}{state code (two letter)}
  \item{zip}{zip code, character format}
}}
\source{
it's made up
}
\usage{
urel
}
\description{
A dataset containing a few UREL regulars.
}
\keyword{datasets}

