% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_management.R
\name{api_key}
\alias{api_key}
\title{Set or get an API key}
\usage{
api_key(provider, force = FALSE)
}
\arguments{
\item{provider}{Which service is this a key for?}

\item{force}{Overwrite existing key? If TRUE, the user will be prompted to 
enter a user key even if one has already been entered before.}
}
\description{
Reads (or writes) an API key to an environment variable, and stores the key 
to disk.
}

