% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Get geo-codes from addresses}
\usage{
geocode(addresses, provider = "google", ...)
}
\arguments{
\item{addresses}{A character vector of addresses}

\item{provider}{Which geo-coding API to use? ("google" or "mapquest")}
}
\value{
A \code{data.frame} with three columns: \code{address}, 
\code{latitude}, and \code{longitude}
}
\description{
Get geo-codes from addresses
}
\seealso{
\code{\link{geocode_google}}, \code{\link{geocode_mapquest}}
}

